refactor(web): rename mis-named draft folders to chrome and components/draft
The styles/draft partials were global yn-* design-system styles, not
draft-specific. Rename the folder to styles/chrome with a chrome.scss
aggregator and move the page-specific neighbours styles onto the
neighbours page.

Relocate the genuine draft-edit toolkit (hooks, reducer, drawer,
action buttons) to components/draft, and move the generic helpers out
from under draft: yaml and row-validation into utils, the create-config
modal into components, and drop the duplicate cidr validator in favour
of the existing utils isValidCidr.
