+++ web/src/components/draft/DraftItemDrawer.tsx
