+++ web/src/components/draft/index.ts
@@ -1,7 +1,5 @@
 export { createDraftReducer } from './draftReducer';
 export type { DraftState, DraftAction, CreateDraftReducerOptions } from './draftReducer';
-export { isValidCIDR } from './cidr';
-export { default as AddConfigModal } from './AddConfigModal';
 export { default as DraftActionButtons, SaveIcon, TrashIcon } from './DraftActionButtons';
 export { default as DraftItemDrawer } from './DraftItemDrawer';
 export { useDraftShortcuts } from './useDraftShortcuts';
@@ -11,8 +9,6 @@ export { useDraftDragDrop } from './useDraftDragDrop';
 export type { DragOverState, UseDraftDragDropResult } from './useDraftDragDrop';
 export { useDraftPageHandlers } from './useDraftPageHandlers';
 export type { UseDraftPageHandlersResult } from './useDraftPageHandlers';
-export { rowHasError, countInvalidRows } from './validation';
-export { dumpYamlDoc, parseYamlList } from './yaml';
 export { DRAWER_TRANSITION_MS } from './constants';
 export { computeRowStatuses } from './rowStatus';
 export type { RowStatusResult } from './rowStatus';
