+++ web/src/pages/modules/acl/AclPage.tsx
@@ -16,12 +16,12 @@ import type { RuleDrawerHandle } from './RuleDrawer';
 import YamlIO, { type ImportMode } from './YamlIO';
 import { SaveDiffModal } from './SaveDiffModal';
 import { useAclRuleCounters } from './useAclRuleCounters';
-import { AddConfigModal, DRAWER_TRANSITION_MS } from '../../_shared/draft';
-import { DeleteConfigModal, BulkDeleteModal, CommandPaletteHeader } from '../../../components';
+import { AddConfigModal, DeleteConfigModal, BulkDeleteModal, CommandPaletteHeader } from '../../../components';
+import { DRAWER_TRANSITION_MS } from '../../../components/draft';
 import { usePalette } from '../../../components/command-palette';
 import type { Command, RowAdapter } from '../../../components/command-palette';
 import { useTabCycle } from '../../_shared/useTabCycle';
-import '../../../styles/draft-page.scss';
+import '../../../styles/chrome.scss';
 import './acl.scss';
 
 const QP_CONFIG = 'config';
