+++ web/src/pages/modules/forward/forward.scss
@@ -3,7 +3,7 @@
 
 // Forward-specific styles — direction badges, sparkline, chip-input, segmented control.
 // Shared page chrome (yn-page, yn-tabs, yn-table-*, yn-modal*, yn-drawer*, yn-bulk-bar, etc.)
-// lives in web/src/styles/draft-page.scss.
+// lives in web/src/styles/chrome.scss.
 
 // Direction badges
 .yn-badge-dir {
