+++ web/src/pages/operators/neighbours/neighbours.scss
@@ -1,4 +1,4 @@
-@use '../mixins' as *;
+@use '../../../styles/mixins' as *;
 
 // Neighbours page extensions.
 // Scoped to .yn-page to avoid bleeding into other pages.
