+++ web/src/styles/chrome/_badges.scss
