+++ web/src/styles/chrome/_surface.scss
