--- web/src/styles/draft-page.scss
@@ -1,12 +1,0 @@
-// Shared page chrome for draft-edit pages (Forward, Route, ACL, Neighbours, etc.).
-// This file is an aggregator of ./draft/* partials — do not add rules here directly.
-// The yn- prefix is historical — these classes are consumed by multiple pages.
-// Design tokens are defined globally in tokens.scss (imported from main.tsx).
-
-@use './draft/surface';
-@use './draft/buttons';
-@use './draft/drawer';
-@use './draft/forms';
-@use './draft/modal';
-@use './draft/badges';
-@use './draft/neighbours';
