+++ web/src/utils/index.ts
@@ -10,5 +10,6 @@ export * from './packetParser';
 export * from './clipboard';
 export * from './counterGroups';
 export * from './validation';
+export * from './yaml';
 export * from './ranges';
 export * from './interpolation';
